only remove maven-lib if it was installed

+++ maybe-replace-maven.sh
@@ -3,7 +3,9 @@
 set -x
 set -e
 if [ -f /files/share/maven/bin/mvn ]; then
-    dnf -y remove maven-lib --noautoremove
+    if rpm -q maven-lib >/dev/null 2>&1; then
+        dnf -y remove maven-lib --noautoremove
+    fi
     cp -R /files/share /usr
     ln -snf /usr/share/maven/bin/mvn /usr/bin
     ls -l /usr/bin/mvn
